+++ backend/interfaces/api/server/handler/command.go
@@ -6,15 +6,15 @@ import (
 	"os/exec"
 )
 
-func (h Handler) Command(w http.ResponseWriter, r *http.Request) error {
+func (h indexHandler) Command(w http.ResponseWriter, r *http.Request) error {
 	cmd := exec.Command("pwd")
 	var out bytes.Buffer
 	cmd.Stdout = &out
 	err := cmd.Run()
 	if err != nil {
 		h.Context.Logger.Fatalln(err)
-		return h.Context.JSON(w, http.StatusOK, err.Error())
+		return h.JSON(w, http.StatusOK, err.Error())
 	}
 
-	return h.Context.JSON(w, http.StatusOK, out.String())
+	return h.JSON(w, http.StatusOK, out.String())
 }
