+++ backend/interfaces/api/server/handler/handler.go
@@ -1,0 +1,7 @@
+package handler
+
+import "github.com/yoshihiro-shu/draft-backend/interfaces/api/server/request"
+
+type Handler struct {
+	*request.Context
+}
